+++ core/src/test/scala/akka/wamp/messages/JsonMessageSpec.scala
@@ -52,12 +52,8 @@ trait JsonMessageBehaviours { this: JsonMessageSpec =>
       val message = s.deserialize(json)
       message match {
         case c: DataConveyor => {
-          whenReady(c.args) { a =>
-            a mustBe empty
-            whenReady(c.kwargs) { a =>
-              a mustBe empty
-            }
-          }
+          c.args mustBe empty
+          c.kwargs mustBe empty
         }
         case msg => fail(s"Unexpected $msg")
       }
@@ -69,11 +65,10 @@ trait JsonMessageBehaviours { this: JsonMessageSpec =>
       val message = s.deserialize(json)
       message match {
         case c: DataConveyor =>
-          whenReady(c.kwargs[UserType]) { user =>
-            user.name mustBe "paolo"
-            user.age mustBe 99
-            user.male mustBe true
-          }
+          val user = c.kwargs[UserType]
+          user.name mustBe "paolo"
+          user.age mustBe 99
+          user.male mustBe true
         case msg =>
           fail(s"Unexpected $msg")
       }
@@ -86,46 +81,44 @@ trait JsonMessageBehaviours { this: JsonMessageSpec =>
       val message = s.deserialize(json)
       message match {
         case c: DataConveyor =>
-          whenReady(c.args) { args =>
-            args must have size (9)
-            // TODO args(0) mustBe (null)
-            args(1) mustBe a[java.lang.Integer]
-            args(1) mustBe 2147483647
-            args(2) mustBe a[java.lang.Long]
-            args(2) mustBe 2147483648L
-            args(3) mustBe a[java.lang.Double]
-            args(3) mustBe 1.45
-            args(4) mustBe a[java.lang.String]
-            args(4) mustBe "string"
-            args(5) mustBe a[java.lang.Boolean]
-            args(5) mustBe true
-            args(6) mustBe a[java.lang.Boolean]
-            args(6) mustBe false
-            args(7) mustBe a[scala.collection.immutable.List[_]]
-            args(7) mustBe List(null, 2147483647, 2147483648L, 1.45, "string", true, false)
-            args(8) mustBe a[scala.collection.immutable.Map[_, _]]
-            args(8) mustBe Map("key0" -> null, "key1" -> 2147483647, "key2" -> 2147483648L, "key3" -> 1.45, "key4" -> "string", "key5" -> true, "key6" -> false)
-          }
-          whenReady(c.kwargs) { kwargs =>
-            kwargs must have size(9)
-            // TODO kwargs("arg0") mustBe null
-            kwargs("arg1") mustBe a[java.lang.Integer]
-            kwargs("arg1") mustBe 2147483647
-            kwargs("arg2") mustBe a[java.lang.Long]
-            kwargs("arg2") mustBe 2147483648L
-            kwargs("arg3") mustBe a[java.lang.Double]
-            kwargs("arg3") mustBe 1.45
-            kwargs("arg4") mustBe a[java.lang.String]
-            kwargs("arg4") mustBe "string"
-            kwargs("arg5") mustBe a[java.lang.Boolean]
-            kwargs("arg5") mustBe true
-            kwargs("arg6") mustBe a[java.lang.Boolean]
-            kwargs("arg6") mustBe false
-            kwargs("arg7") mustBe a[scala.collection.immutable.List[_]]
-            kwargs("arg7") mustBe List(null,2147483647,2147483648L,1.45,"string",true,false)
-            kwargs("arg8") mustBe a[scala.collection.immutable.Map[_, _]]
-            kwargs("arg8") mustBe Map("key0"->null,"key1"->2147483647,"key2"->2147483648L,"key3"->1.45,"key4"->"string","key5"->true,"key6"->false)
-          }
+          c.args must have size (9)
+          // TODO c.args(0) mustBe (null)
+          c.args(1) mustBe a[java.lang.Integer]
+          c.args(1) mustBe 2147483647
+          c.args(2) mustBe a[java.lang.Long]
+          c.args(2) mustBe 2147483648L
+          c.args(3) mustBe a[java.lang.Double]
+          c.args(3) mustBe 1.45
+          c.args(4) mustBe a[java.lang.String]
+          c.args(4) mustBe "string"
+          c.args(5) mustBe a[java.lang.Boolean]
+          c.args(5) mustBe true
+          c.args(6) mustBe a[java.lang.Boolean]
+          c.args(6) mustBe false
+          c.args(7) mustBe a[scala.collection.immutable.List[_]]
+          c.args(7) mustBe List(null, 2147483647, 2147483648L, 1.45, "string", true, false)
+          c.args(8) mustBe a[scala.collection.immutable.Map[_, _]]
+          c.args(8) mustBe Map("key0" -> null, "key1" -> 2147483647, "key2" -> 2147483648L, "key3" -> 1.45, "key4" -> "string", "key5" -> true, "key6" -> false)
+
+          c.kwargs must have size(9)
+          // TODO c.kwargs("arg0") mustBe null
+          c.kwargs("arg1") mustBe a[java.lang.Integer]
+          c.kwargs("arg1") mustBe 2147483647
+          c.kwargs("arg2") mustBe a[java.lang.Long]
+          c.kwargs("arg2") mustBe 2147483648L
+          c.kwargs("arg3") mustBe a[java.lang.Double]
+          c.kwargs("arg3") mustBe 1.45
+          c.kwargs("arg4") mustBe a[java.lang.String]
+          c.kwargs("arg4") mustBe "string"
+          c.kwargs("arg5") mustBe a[java.lang.Boolean]
+          c.kwargs("arg5") mustBe true
+          c.kwargs("arg6") mustBe a[java.lang.Boolean]
+          c.kwargs("arg6") mustBe false
+          c.kwargs("arg7") mustBe a[scala.collection.immutable.List[_]]
+          c.kwargs("arg7") mustBe List(null,2147483647,2147483648L,1.45,"string",true,false)
+          c.kwargs("arg8") mustBe a[scala.collection.immutable.Map[_, _]]
+          c.kwargs("arg8") mustBe Map("key0"->null,"key1"->2147483647,"key2"->2147483648L,"key3"->1.45,"key4"->"string","key5"->true,"key6"->false)
+
         case msg =>
           fail(s"Unexpected $msg")
       }