+++ examples/futures/src/main/java/FuturesJavaClient.java
@@ -3,10 +3,17 @@
  * file 'LICENSE.md', which is part of this source code package.
  */
 
-import akka.Done;
+/*
+ * This file is subject to the terms and conditions defined in
+ * file 'LICENSE.md', which is part of this source code package.
+ */
+
 import akka.actor.ActorSystem;
 import akka.wamp.client.japi.*;
 import com.typesafe.config.*;
+
+import java.util.List;
+import java.util.concurrent.CompletableFuture;
 
 import static java.util.Arrays.asList;
 import static java.lang.System.out;
@@ -24,25 +31,20 @@ public class FuturesJavaClient {
         s.publish("topic", asList("Ciao!"));
 
         s.subscribe("topic", (event) -> {
-          return event.args().thenApply(args -> {
-            out.println("got " + args.get(0));
-            return Done.getInstance();
-          });
+          out.println("got " + event.args().get(0));
         });
 
         s.register("procedure", (invocation) -> {
-          return invocation.args().thenApply(args -> {
-            Integer a = (Integer) args.get(0);
-            Integer b = (Integer) args.get(1);
-            Integer result = a + b;
-            return Payload.create(asList(result));
-          });
+          List<Object> args = invocation.args();
+          Integer a = (Integer) args.get(0);
+          Integer b = (Integer) args.get(1);
+          Integer result = a + b;
+          // TODO Simplify Java API
+          return CompletableFuture.completedFuture(Payload.create(asList(result)));
         });
 
         s.call("procedure", asList(20, 55)).thenAccept(res -> {
-          res.args().thenAccept(args -> {
-            out.println("20 * 55 = " + args.get(0));
-          });
+          out.println("20 * 55 = " + res.args().get(0));
         });
       });
     });