+++ examples/poloniex/src/main/java/PoloniexJavaClient.java
@@ -1,19 +1,16 @@
-
-import akka.Done;
-import akka.actor.ActorSystem;
-import akka.wamp.client.japi.Client;
+import akka.actor.*;
+import akka.wamp.client.japi.*;
 
 public class PoloniexJavaClient {
   public static void main(String[] aaa) {
     ActorSystem actorSystem = ActorSystem.create();
     Client client = Client.create(actorSystem);
-    client.connect("wss://api.poloniex.com", "json").thenAccept(c -> {
-      c.open("realm1").thenAccept(s -> {
-        s.subscribe("ticker", event -> {
-          return event.args().thenApply(args -> {
-              System.out.println(args.toString());
-              return Done.getInstance();
-            });
+
+    client.connect("wss://api.poloniex.com", "json").thenAccept(conn -> {
+      conn.open("realm1").thenAccept(session -> {
+
+        session.subscribe("BTC_XMR", event -> {
+          System.out.printf("%s --> %s\n", event.kwargs(), event.args());
         });
       });
     });